submit button CSS
changed bottom element to a container element containing "your answers", rather than just "yourAnswers" being the bottom element. 
Broke up "yourAnswersContainer" into a CSS grid with 1 row (10% of the height of the super container, i.e. the whole height of the container available, 1fr) and 2 columns (both set at 1fr). "buttonContainer" span is itself only as tall as the button, is moved down vertically as far as possible within its grid space (justify-self: end) and the button it contains was moved to the right with "text-align: right"